feat: 创建管理任务的接口

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/entity/commontask/CommonTask.java
@@ -195,7 +195,7 @@ public class CommonTask extends BaseEntity {
      * 父级id
      */
     @Column(name = "parent_id")
-    private Integer parentId;
+    private Long parentId;
 
     @Transient
     private List<TaskProduct> productList;
